requirements.txt
hiredis==3.0.0
redis==5.0.7
python-dotenv==1.0.1
requests==2.32.3
kiteconnect==5.0.1
pandas==2.2.3
pyotp==2.9.0
PyQt5==5.15.11